requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
argcomplete==1.10.3
attrs==25.3.0
azure-ai-documentintelligence==1.0.0b4
backoff==2.2.1
beautifulsoup4==4.8.2
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
chromadb==0.5.0
click==8.2.1
colorama==0.4.6
compressed-rtf==1.0.7
cryptography==45.0.3
dataclasses-json==0.6.7
distro==1.9.0
docx2txt==0.9
ebcdic==1.1.1
emoji==2.14.1
et-xmlfile==2.0.0
extract-msg==0.28.7
filetype==1.2.0
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.73.0rc1
grpcio-status==1.73.0rc1
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
IMAPClient==2.1.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.62
langchain-google-genai==2.1.5
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langdetect==1.0.9
langsmith==0.3.42
lxml==5.4.0
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.41.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.2
olefile==0.47
openai==1.82.0
openpyxl==3.1.5
orjson==3.10.18
packaging==23.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20191110
pgvector==0.2.4
pillow==11.2.1
plotly==6.1.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.0
psutil==7.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.9
pyarrow==20.0.0
pyvis==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
PyMuPDF==1.26.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-iso639==2025.2.18
python-magic==0.4.27
python-oxmsg==0.0.2
python-pptx==0.6.23
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.25.1
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.14.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.8.1
SQLAlchemy==2.0.23
streamlit==1.45.1
sudachipy==0.6.8
sudachidict-core==20250129
tenacity==8.5.0
tiktoken==0.9.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
watchdog==6.0.0
webencodings==0.5.1
wrapt==1.17.2
xlrd==1.2.0
XlsxWriter==3.2.3
yarl==1.20.0
zstandard==0.23.0